sending invite api

+++ onboard-service/src/main/java/com/zinios/onboard/Repository/OtpRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface OtpRepository extends JpaRepository<Otp, Integer> {
+public interface OtpRepository extends JpaRepository<Otp, Long> {
     Optional<Otp> findTopByEmailAndOtpAndIsActiveTrueOrderByCreatedTimeDesc(String email, String otp);
 
     @Transactional
